Put Stoke data getters in analysis folder getter file
